requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==0.23
more-itertools==8.0.0
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
python-decouple==3.3
requests==2.22.0
six==1.13.0
spotipy==2.4.4
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
